requirements.txt
Authlib==1.3.2
aiosqlite==0.20.0
certifi==2024.8.30
cffi==1.17.1
click==8.1.7
cryptography==43.0.3
fastapi==0.115.4
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
idna==3.10
itsdangerous==2.2.0
pycparser==2.22
pydantic==2.9.2
PyJWT==2.9.0
python-multipart==0.0.16
rfc3986==2.0.0
sniffio==1.3.1
starlette==0.41.2
SQLAlchemy==2.0.36
typing-extensions==4.12.2
uvicorn==0.32.0
python-dotenv==1.0.1